+++ changelog.md
@@ -2,6 +2,11 @@
 
 All notable changes to the BSPED Paediatric DKA Calculator codebase will be documented in this file.
 
+See also:
+
+- [API repo](https://github.com/dan-leach/dka-calculator-api)
+- [contributors](https://github.com/dan-leach/dka-calculator/blob/main/contributors.md) to the project.
+
 ## [v2.0] - TBC 2024
 
 ### Changed
@@ -21,6 +26,12 @@ All notable changes to the BSPED Paediatric DKA Calculator codebase will be docu
 - The insulin delivery method the patient uses (if they have pre-existing diabetes) is now collected for audit purposes.
 - Factors that may have contributed to the episode of DKA occuring (preventable factors) are now collected for audit purposes.
 - When the patient has pre-existing diabetes ticks are automatically added to the not applicable boxes for ongoing insulin therapy for the mode the patient does not use on flow chart 5.
+
+## [v1.3.6] - 2024-03-12 16:00
+
+## Changed
+
+- Security update to fix SQL injection vulnerabilities
 
 ## [v1.3.5] - 2022-12-06 08:55